+++ src/ui/assets.d.ts
@@ -11,3 +11,8 @@ declare module '*.png' {
   const src: string;
   export default src;
 }
+
+declare module '*.webp' {
+  const src: string;
+  export default src;
+}